+++ lib/elixir_script.ex
@@ -3,6 +3,7 @@ defmodule ElixirScript do
   alias ElixirScript.Translator.JSModule
   alias ESTree.Tools.Builder
   alias ESTree.Tools.Generator
+  require Logger
 
   @moduledoc """
   Transpiles Elixir into JavaScript.
@@ -15,10 +16,12 @@ defmodule ElixirScript do
   that controls transpiler output.
 
   Available options are:
-  * include_path: a boolean controlling whether to return just the JavaScript code
+  * `:include_path` - a boolean controlling whether to return just the JavaScript code
   or a tuple of the file name and the JavaScript code
 
-  * root: a binary path prepended to the path of the standard lib imports if needed
+  * `:root` - a binary path prepended to the path of the standard lib imports if needed
+  * `:env` - a Macro.env struct to use. This is most useful when using macros. Make sure that the
+  given env has the macros required. Defaults to __ENV__.
   """
 
   @doc """
@@ -38,8 +41,9 @@ defmodule ElixirScript do
   def transpile_quoted(quoted, opts \\ []) do
     include_path = Dict.get(opts, :include_path, false)
     root = Dict.get(opts, :root)
+    env = Dict.get(opts, :env, __ENV__)
 
-    case Translator.translate(quoted) do
+    case Translator.translate(quoted, env) do
       modules when is_list(modules) ->
         List.flatten(modules)
         |> Enum.map(fn(x) ->
@@ -59,13 +63,14 @@ defmodule ElixirScript do
   def transpile_path(path, opts \\ []) do
     include_path = Dict.get(opts, :include_path, false)
     root = Dict.get(opts, :root)
+    env = Dict.get(opts, :env, __ENV__)
 
     path
     |> Path.wildcard
     |> Enum.map(fn(x) -> 
       File.read!(x)
       |> Code.string_to_quoted!
-      |> Translator.translate
+      |> Translator.translate(env)
     end)
     |> List.flatten
     |> Enum.map(fn(x) ->