+++ lib/elixir_script/translator/assignment.ex
@@ -1,11 +1,10 @@
 defmodule ElixirScript.Translator.Assignment do
   @moduledoc false
-  require Logger
   alias ESTree.Tools.Builder, as: JS
   alias ElixirScript.Translator
-  alias ElixirScript.Translator.PatternMatching.Match
+  alias ElixirScript.PatternMatching.Match
 
-  def make_assignment(left, right) do
+  def make_assignment(left, right, env) do
     { patterns, params } = Match.build_match([left])
 
       declarator = JS.variable_declarator(
@@ -15,7 +14,7 @@ defmodule ElixirScript.Translator.Assignment do
             JS.identifier("fun"),
             JS.identifier("bind")
           ),
-          [hd(patterns), Translator.translate(right)]
+          [hd(patterns), Translator.translate(right, env)]
         )
       )