+++ lib/elixir_script/translator/case.ex
@@ -1,16 +1,16 @@
 defmodule ElixirScript.Translator.Case do
   @moduledoc false
-  alias ESTree.Tools.Builder
+  alias ESTree.Tools.Builder, as: JS
   alias ElixirScript.Translator
   alias ElixirScript.Translator.Function
   
-  def make_case(condition, clauses) do
-    Builder.call_expression(
-      Builder.member_expression(
-        Function.make_anonymous_function(clauses),
-        Builder.identifier("call")
+  def make_case(condition, clauses, env) do
+    JS.call_expression(
+      JS.member_expression(
+        Function.make_anonymous_function(clauses, env),
+        JS.identifier("call")
       ),
-      [Translator.translate(condition)]
+      [JS.identifier(:this), Translator.translate(condition)]
     )
   end  
 end