+++ lib/elixir_script/translator/receive.ex
@@ -1,10 +1,9 @@
 defmodule ElixirScript.Translator.Receive do
   @moduledoc false
-  require Logger
   alias ESTree.Tools.Builder, as: JS
   alias ElixirScript.Translator
 
-  def make_receive([do: clauses]) do
+  def make_receive([do: clauses], env) do
     JS.call_expression(
       JS.member_expression(
         JS.member_expression(
@@ -19,7 +18,7 @@ defmodule ElixirScript.Translator.Receive do
           [],
           JS.block_statement([
             JS.return_statement(
-              ElixirScript.Translator.Case.make_case({:__aliases__, [], [:message]}, clauses)
+              ElixirScript.Translator.Case.make_case({:__aliases__, [], [:message]}, clauses, env)
             )
           ])
         )
@@ -27,7 +26,7 @@ defmodule ElixirScript.Translator.Receive do
     )
   end
 
-  def make_receive([do: clauses, after: [{:->, _, [[time], _body]}] = after_clause]) do
+  def make_receive([do: clauses, after: [{:->, _, [[time], _body]}] = after_clause], env) do
     JS.call_expression(
       JS.member_expression(
         JS.member_expression(
@@ -42,12 +41,12 @@ defmodule ElixirScript.Translator.Receive do
           [],
           JS.block_statement([
             JS.return_statement(
-              ElixirScript.Translator.Case.make_case({:__aliases__, [], [:message]}, clauses)
+              ElixirScript.Translator.Case.make_case({:__aliases__, [], [:message]}, clauses, env)
             )
           ])
         ),
-        Translator.translate(time),
-        ElixirScript.Translator.Function.make_anonymous_function(after_clause)    
+        Translator.translate(time, env),
+        ElixirScript.Translator.Function.make_anonymous_function(after_clause, env)    
       ]
     ) 
   end