+++ lib/elixir_script/translator/struct.ex
@@ -1,6 +1,5 @@
 defmodule ElixirScript.Translator.Struct do
   @moduledoc false
-  require Logger
   alias ESTree.Tools.Builder, as: JS
   alias ElixirScript.Translator